PINN_Temporal_Graph/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
scipy>=1.7.0
pyyaml>=5.4.0
torch-geometric>=2.0.0
tqdm>=4.62.0